New approach for playlists
It appears youtube has killed the old endpoint feeding us JSON directly.
Proceed to extracting it manually from the initial HTML document.

The JSON in that script tag is not exactly fun to play with, use a
hybrid approach with go-simplejson to keep sanity even if it might not
be the most efficient.