Stop favoring format 137
I made a PR that was merged:

https://github.com/kkdai/youtube/pull/170

but the issue is now, that my fix works for all formats, except one:

    +------+-----+---------+---------+----------+-----------+---------+---------
    | ITAG | FPS |  VIDEO  |  AUDIO  |  AUDIO   | SIZE [MB] | BITRATE |
    |      |     | QUALITY | QUALITY | CHANNELS |           |         |
    +------+-----+---------+---------+----------+-----------+---------+---------
    |  137 |  25 | 1080p   |         |        0 |      16.2 |  366467 | video/mp
    +------+-----+---------+---------+----------+-----------+---------+---------

For some reason, this format still goes slow, even with the fix. We could
implement concurrent downloader to fix it, but I think a better option is to
just leave the download code as is, but change the default picker.